travis-ci: apt-get install ant-optional
travis suddenly did not work anymore with the following error:

test:
    [mkdir] Created dir: /home/travis/build/exedio/persistence-console/build/testresults

BUILD FAILED
/home/travis/build/exedio/persistence-console/build.xml:119: The following error occurred while executing this line:
/home/travis/build/exedio/persistence-console/build.xml:32: Problem: failed to create task or type junit
Cause: the class org.apache.tools.ant.taskdefs.optional.junit.JUnitTask was not found.
        This looks like one of Ant's optional components.
Action: Check that the appropriate optional JAR exists in
        -/usr/share/ant/lib
        -/home/travis/.ant/lib
        -a directory added on the command line with the -lib argument

Do not panic, this is a common problem.
The commonest cause is a missing JAR.

This is not a bug; it is a configuration problem